[patch] bugfix of windows environment
