+++ now_windows.go
@@ -1,0 +1,9 @@
+// +build windows
+
+package fastime
+
+import "time"
+
+func (f *Fastime) now() time.Time {
+	return time.Now()
+}
